+++ src/app/app.tsx
@@ -1,4 +1,4 @@
-import MainPage from '../pages/main-page';
+import MainPage from '../pages/main-page/main-page';
 
 type AppProps = {
   placesCount: number;
