+++ src/components/place-card/place-card.tsx
@@ -1,4 +1,4 @@
-function PlaceCard (): JSX.Element {
+export function PlaceCard(): JSX.Element {
   return (
     <article className="cities__card place-card">
       <div className="place-card__mark">
@@ -42,5 +42,3 @@ function PlaceCard (): JSX.Element {
     </article>
   );
 }
-
-export default PlaceCard;
